Structural Changes
* Add internal_buildsys

Enable the use of an "internal buildsys" to enable lazybuilder to be
part of a more intricate build system if at all necessary for the user.

* Module dist tracker

Add a file to track the dist changes (changing from a 000 placeholder)

* Add solo support

mock.cfg.j2 was missing solo_repo support

* Restructure mock_vars.yml

Some vars were being called out before being defined. This is especially
true for bootstrap or solo builds. This is an attempt to resolve the
issue.

* Add ETI issue templates

For project owners (@nazunalika), this is a simple template for our own
tracking in github or gitlab. A user of the project should not choose
this template when opening an issue.